pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.29",
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "djoser>=2.3.1",
    "gunicorn>=23.0.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
]
